fix(3nf): W1 cache double precision + extend key with c1/c3/c4/grid (B5+B6)
Phase 6 of 3NF audit (docs/3nf_audit_2026-06-21.md):

B5: W1Key extended with c_1, c_3, c_4 (GeV^-1) and SHA-256 hashes of
p_WP_array / q_WP_array. W1_SCHEMA_VERSION bumped 3 -> 4 so old caches
read as misses. Prevents wrong reuse across Hamiltonians or grids that
share c_D/c_E/Lambda but differ in 2NF c_i (e.g. N2LOopt vs Idaho_N3LO).

B6: W1 cache storage changed from std::vector<float> to std::vector<double>.
3NF has strong cancellations between c_E/c_D/c_1/c_3 contributions; float
introduces ~1e-5 relative noise that compounds under Pade. Double removes
this noise floor.

LEC accessors lec_c1_gev/lec_c3_gev/lec_c4_gev added to chiral_N2LO_3NF
and three_nucleon_force_model base so the cache layer can read them.

sha256: add hex_digest_raw for hashing raw byte arrays (grid hashing).
test: cache_layer_test updated for new schema + double storage.

+++ src/interactions/w1_pw_cache.h
@@ -43,7 +43,7 @@ public:
         if (blk < 0) return 0.0;
         const std::size_t idx =
             ((((std::size_t)blk * m_Nq + idx_q_r) * m_Nq + idx_q_c) * m_Np + idx_p_r) * m_Np + idx_p_c;
-        return (double)m_data[idx];
+        return m_data[idx];
     }
 
     bool        valid()       const { return !m_data.empty(); }
@@ -51,10 +51,15 @@ public:
     std::size_t Nq()          const { return m_Nq; }
     std::size_t Np()          const { return m_Np; }
     std::size_t num_blocks()  const { return m_blocks.size(); }
-    std::size_t total_bytes() const { return m_data.size() * sizeof(float); }
+    // 3NF audit B6 (2026-06-21): storage is now double (was float).
+    // 24-bit float mantissa was insufficient for cancelling 3NF contributions.
+    std::size_t total_bytes() const { return m_data.size() * sizeof(double); }
 
 private:
-    std::vector<float>                   m_data;
+    // 3NF audit B6: changed from std::vector<float> to std::vector<double>.
+    // For memory-constrained runs, the previous float storage can be restored
+    // behind an explicit compile-time flag (not currently exposed).
+    std::vector<double>                  m_data;
     std::vector<int>                     m_block_index;  // (alpha_r, alpha_c) → block id, -1 if forbidden
     std::vector<std::pair<int, int>>     m_blocks;       // block id → (alpha_r, alpha_c)
     std::size_t                          m_Nalpha = 0;
