+++ src/io/cache_layer/cache_io_w1.h
@@ -6,11 +6,12 @@
 namespace tictac::cache {
 
 struct W1Block {
-    int                Nq;        // q-WP count for this block
-    int                Np;        // p-WP count for this block
-    int                a_r;       // alpha row index (informational)
-    int                a_c;       // alpha col index (informational)
-    std::vector<float> data;      // size = Nq*Nq*Np*Np
+    int                 Nq;        // q-WP count for this block
+    int                 Np;        // p-WP count for this block
+    int                 a_r;       // alpha row index (informational)
+    int                 a_c;       // alpha col index (informational)
+    // 3NF audit B6 (2026-06-21): storage changed from float to double.
+    std::vector<double> data;      // size = Nq*Nq*Np*Np
 };
 
 bool read_w1_h5 (const std::string& path,
